requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Flask==1.1.4
Flask-Migrate==3.0.1
flask-restx==0.4.0
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
livereload==2.6.3
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.17
tornado==6.1
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
